test playerfactory method
