--- Test/script.test.js
@@ -1,3 +1,0 @@
-import {playerFactory, board, playGame} from '../js/script'
-
-describe("Player factor")
