+++ index.html
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Tic-Tac-Toe Game</title>
     <link rel="stylesheet" href="styles.css" />
-    <script src="js/script.js" defer></script>
+    <script type="module" src="./js/playGame.js" defer></script>
   </head>
   <body>
     <h1 class="heading">Tic Tac Toe</h1>
