+++ js/playerFactory.js
@@ -1,0 +1,3 @@
+const playerFactory = (name, symbol) => ({ name, symbol });
+
+export default playerFactory
